sum2table.pl: Better support for handling a single sum file
Previously when handling a single sum file, every line had the '(-)'
suffix added to it, indicating the test had disappeared from the second
set of results.  This is now fixed, and for a single summary file, no
suffix is added.

Sorting is also improved.  When using the 'BAD' specifier with a single
summary file we now sort on absolute result value, which places all of
the test scripts with failing results at the top of the table.