compare-tests: New script to compare two summary files
This is a new script to compare two DeJaGnu summary files and report any
regressions.  This is similar in spirit to the script that can be found
within the GCC source tree, however, I've found that script seems to
miss-identify test names, and lacks the ability to focus the output onto
"bad" events (regressions).