libs: Improve SumFileParser
Ran into some .sum files that didn't have any 'Running' lines at all, I
needed to tweak the SumFileParser to allow it to parse these files.

The idea now is that if we see a test status before a 'Running' line,
then we expect to never see a 'Running' line.  If we see a test status
without an exp filename prefix, then we expect to have seen a 'Running'
line.