sum2table: Don't add whitespace suffix unless really needed
When displaying lines in the results table and comparing two summary
files, we would previously always add a whitespace suffix.  This was
to allow for the case where a '(+)' or '(-)' might need to be
displayed.

With this commit if we know that no '(+)' or '(-)' will be displayed
then drop the whitespace suffix, this makes the table look a little
neater.