Split test name filtering into a separate module
This is the start of an attempt to cleanup how test name filtering is
managed.

This first step splits the filtering into a separate module, and puts
in place a framework for having tool specific filters.  At this point
there are no tool specific filters added, this will hopefully come
along later.