sum-duplicates: New script to find duplicate test names
New tool 'sum-duplicates' finds duplicate test names.  Along with this
I've added a mechanism so that it is possible to load a .sum file
without filtering the test names, however, in the end I have taken the
approach of just adding a new method to TestResult so that we can get
the unfiltered test name.  I've left in the option of not loading the
filters though as this is a tiny change, and could be useful one day.