Support for fetching results for specific targets
The SumFile object now supports fetching results for a specific
target, or for all targets.  The TestResult objects can now be asked
which target they are for.

After this commit all of the tools will once again be operating on the
complete set of results from all targets, as we don't take target
names into account when matching results this will mean a result from
one target will happily match a result from a different target.