Update all tools to handle multiple targets in a single summary file
This commit updates all of the tools to handle multiple targets within
a single summary file.

Not all of the tools really handle multiple targets in the best
possible way, the sum-duplicates tools does a reasonable job, printing
all the results for all targets in one go, but all of the other tools
require the user to specify which target to print results for, after
that the tools just selects and handles the results for that one
target.

In future I'd like to update more of the tools to handle all targets
in one go, but for now this will do.